+++ CHANGELOG.md
@@ -6,9 +6,13 @@
 ### 0.06.26001
 1. **Add feature**: Add Loading Screen when Sinote is loading plugin or loading font.
 2. **Fix issue**: Fix issue ``When trigger Save As, it will be pop out a Exception Window``
+3. **Add feature in CLI**: Realized ``check`` command for check plugin
+4. **Add feature**: Add many shortcuts to Sinote.
+5. **Add feature**: Add Pop-up Dialog to confirm exit Sinote.
+6. **Fix issue**: Replace Python normal `dict` to `SafetyDict` for safe.
 ### 0.06.25542
 1. **Add feature**: Add **+** button to create new Tab. (I forgot add it in 0.06.25530)
-2. **Add feature**: It automatically restores last session every start. (It won't be restore your cursor position.)
+2. **Add feature**: It automatically restores last session every start. (It won't be restored your cursor position.)
 3. **Re-write**: Object *SpacingSupportEdit* **re-write** and discard before version. **(Before was written by AI)**
 4. **Add feature**: Support **Fallback Font**. (Normal will be ***MiSans VF***)
 5. **Add feature**: Support **Theme change**. (It's different from FluentCpp, it can change, but it need not to re-start application.)